requirement.txt
Django==5.1.5
django-rest-framework==0.1.0
djangorestframework==3.15.2
drf-yasg==1.21.8
psycopg2==2.9.10